[Container] Move Data class to Data namespace.
