+++ src/Valkyrja/Container/Manager/Container.php
@@ -15,7 +15,7 @@ namespace Valkyrja\Container\Manager;
 
 use Override;
 use Valkyrja\Container\Contract\Service;
-use Valkyrja\Container\Data;
+use Valkyrja\Container\Data\Data;
 use Valkyrja\Container\Exception\InvalidArgumentException;
 use Valkyrja\Container\Manager\Contract\Container as Contract;
