+++ tests/Unit/Application/Entry/AppTest.php
@@ -24,7 +24,7 @@ use Valkyrja\Cli\Interaction\Output\Output;
 use Valkyrja\Cli\Routing\Collection\Contract\Collection as CliCollection;
 use Valkyrja\Cli\Routing\Data\Data as CliData;
 use Valkyrja\Cli\Routing\Data\Route;
-use Valkyrja\Container\Data as ContainerData;
+use Valkyrja\Container\Data\Data as ContainerData;
 use Valkyrja\Dispatcher\Data\MethodDispatch;
 use Valkyrja\Event\Data as EventData;
 use Valkyrja\Http\Message\Response\Response;
